+++ src/Pages/WellMaster/AddWell/AddWell.jsx
@@ -169,7 +169,6 @@ function AddWell() {
         </Grid>
       </Paper>
 
-
       <Grid
         container
         sx={{
